Move service strategy to registry/service package
