--- pkg/api/rest/create_test.go
@@ -1,41 +1,0 @@
-/*
-Copyright 2014 The Kubernetes Authors All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package rest
-
-import (
-	"testing"
-
-	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/errors"
-)
-
-func TestCheckGeneratedNameError(t *testing.T) {
-	expect := errors.NewNotFound("foo", "bar")
-	if err := CheckGeneratedNameError(Services, expect, &api.Pod{}); err != expect {
-		t.Errorf("NotFoundError should be ignored: %v", err)
-	}
-
-	expect = errors.NewAlreadyExists("foo", "bar")
-	if err := CheckGeneratedNameError(Services, expect, &api.Pod{}); err != expect {
-		t.Errorf("AlreadyExists should be returned when no GenerateName field: %v", err)
-	}
-
-	expect = errors.NewAlreadyExists("foo", "bar")
-	if err := CheckGeneratedNameError(Services, expect, &api.Pod{ObjectMeta: api.ObjectMeta{GenerateName: "foo"}}); err == nil || !errors.IsServerTimeout(err) {
-		t.Errorf("expected try again later error: %v", err)
-	}
-}
