+++ pkg/api/rest/types.go
@@ -17,10 +17,7 @@ limitations under the License.
 package rest
 
 import (
-	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/validation"
 	"k8s.io/kubernetes/pkg/runtime"
-	"k8s.io/kubernetes/pkg/util/fielderrors"
 )
 
 // ObjectFunc is a function to act on a given object. An error may be returned
@@ -42,52 +39,4 @@ func AllFuncs(fns ...ObjectFunc) ObjectFunc {
 		}
 		return nil
 	}
-}
-
-// svcStrategy implements behavior for Services
-// TODO: move to a service specific package.
-type svcStrategy struct {
-	runtime.ObjectTyper
-	api.NameGenerator
-}
-
-// Services is the default logic that applies when creating and updating Service
-// objects.
-var Services = svcStrategy{api.Scheme, api.SimpleNameGenerator}
-
-// NamespaceScoped is true for services.
-func (svcStrategy) NamespaceScoped() bool {
-	return true
-}
-
-// PrepareForCreate clears fields that are not allowed to be set by end users on creation.
-func (svcStrategy) PrepareForCreate(obj runtime.Object) {
-	service := obj.(*api.Service)
-	service.Status = api.ServiceStatus{}
-}
-
-// PrepareForUpdate clears fields that are not allowed to be set by end users on update.
-func (svcStrategy) PrepareForUpdate(obj, old runtime.Object) {
-	// TODO: once service has a status sub-resource we can enable this.
-	//newService := obj.(*api.Service)
-	//oldService := old.(*api.Service)
-	//newService.Status = oldService.Status
-}
-
-// Validate validates a new service.
-func (svcStrategy) Validate(ctx api.Context, obj runtime.Object) fielderrors.ValidationErrorList {
-	service := obj.(*api.Service)
-	return validation.ValidateService(service)
-}
-
-func (svcStrategy) AllowCreateOnUpdate() bool {
-	return true
-}
-
-func (svcStrategy) ValidateUpdate(ctx api.Context, obj, old runtime.Object) fielderrors.ValidationErrorList {
-	return validation.ValidateServiceUpdate(old.(*api.Service), obj.(*api.Service))
-}
-
-func (svcStrategy) AllowUnconditionalUpdate() bool {
-	return true
 }
